pyproject.toml
[tool.poetry]
name = "kg-idg"
version = "0.0.2"
description = "Knowledge Graph for Illuminating the Druggable Genome"
authors = ["caufieldjh <j.harry.caufield@gmail.com>"]
keywords = ["knowledge graph", "bioinformatics"]

[tool.poetry.dependencies]
python = "^3.9"
biolink-model = "^3.0.3"
click = "^8.1.3"
grape = "^0.1.27"
kghub-downloader = "^0.3.1"
kgx = "^1.7.2"
koza = "^0.2.1"
multi-indexer = "^0.0.5"
mysql-connector-python = "^8.0.31"
psycopg2-binary = "^2.9.5"
PyYAML = ">=5.3.1,<6.0.0"
tqdm = "^4.64.1"

[tool.poetry.dev-dependencies]
pytest = "^7.2.0"
parameterized = "^0.8.1"

[tool.black]
line-length = 100
target-version = ["py39", "py310"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
reverse_relative = true